refactor(front): Detach component `App` from `index.jsx`
The structure of the project is now clearer.
