+++ front/src/component/styles/App.module.css
@@ -1,0 +1,5 @@
+.App {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+}
