+++ front/src/index.jsx
@@ -3,65 +3,24 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { ApolloProvider } from '@apollo/client';
 import { ThemeProvider } from 'theme-ui';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 
 import store from './store/store';
 import { client } from './client';
 import theme from './util/theme';
 
-// Import pages.
-import Index from './page/Index';
-import _404 from './page/404';
-import Info from './page/Info';
-import NewsPage from './page/News';
-import SignIn, { URL as signInUrl } from './page/SignIn';
-import Resgister, { URL as resgisterUrl } from './page/Register';
-import PostNews, { URL as postNewsUrl } from './page/PostNews';
+import App from './component/App';
 
-// Import components.
-import Nav from './component/Nav/Nav';
-import Footer from './component/Footer';
-import ScrollToTop from './component/ScrollToTop.tsx';
-import Preflight from './component/Preflight';
+// Import styles.
+import './index.scss';
+
+import appStyles from './component/styles/App.module.css';
 
 if (process.env.NODE_ENV === 'development') {
   const { worker } = await import('./mock/browser');
   worker.start();
   console.log('Add mock...');
 }
-
-const App = () => {
-  return (
-    <React.Fragment>
-      <Preflight />
-      <ScrollToTop />
-      <div
-        sx={{
-          color: 'text',
-          bg: 'background',
-          display: 'flex',
-          flexDirection: 'column',
-          minHeight: '100vh',
-        }}
-      >
-        <Nav />
-        <Routes>
-          <Route path="/" element={<Index />} />
-          <Route path="info" element={<Info />} />
-          <Route path="news">
-            <Route index element={<_404 />} />
-            <Route path=":newsId" element={<NewsPage />} />
-          </Route>
-          <Route path={signInUrl} element={<SignIn />} />
-          <Route path={resgisterUrl} element={<Resgister />} />
-          <Route path={postNewsUrl} element={<PostNews />} />
-          <Route path="*" element={<_404 />} />
-        </Routes>
-        <Footer />
-      </div>
-    </React.Fragment>
-  );
-};
 
 ReactDOM.render(
   <React.StrictMode>
@@ -77,6 +36,8 @@ ReactDOM.render(
   </React.StrictMode>,
   document.getElementById('root'),
 );
+
+document.getElementById('root').classList.add(appStyles.App);
 
 // Hot Module Replacement(HMR)
 if (import.meta.hot) {
